tools/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-genai==1.14.0
python-dotenv==1.0.1
Pillow==11.1.0
python-multipart==0.0.20
